Added conditional to Makefile
In order to control the etoolbox toggle in the the latex I defined a
macro and then used \input when calling pdflatex. This way I can check
for that macro in the latex and modify the build based on it's value or
in this case just whether or not it is defined. This was descibed here:

http://tex.stackexchange.com/questions/1492/passing-parameters-to-a-document